dev (cmd - sidecars): Kafka sidecar new initialization
